Adds option for custom redirect on all routes for those not logged in; defaults to login. Also added frontpage.
